Allow entering of experience scores
Retrieving is currently zero-optimized and we are going to make stuff
up (for the tree) to render it efficiently.  However, not all details
will be shown at all times, and the tree can be fetched in an
intelligent manner.

We should be able to consume a bunch of settings from the hindrances
that we find in the Excel and how these are hooked together.