Data entry for the bulk of the current Building/Entrance
The section building/entrance is not huge, but it does represent an
important milestone as it's part of the currently finalized set of
elements we're consuming.

Note that this is a starting point, and there are a few things which
will need updating.  The commit also did a large sweep of added
contents and compnoents.  Some might not be needed.