A first PoC to show we can render trees with experiences
The current implementation overrides the name of a node.  I doubt that
is the prettiest way to handle this case and I think we should
consider an alternative approach in which we copy-over or abstract the
state machine which handles the opening, closing and loading of the
parts of the tree.

I did not check if there already is some code duplication or not.  If
it is currently there it would make sense to further abstract.