Construct documentation for production environment for now
We should figure out a nice way to render the documentation for either
the QA environment or for the production environment.  From a
distance, that should be easy to do.  In the light of quickly
supplying documentation to the relevant parties, we're opting for
hardcoding the documentation to the produciton environment.