Only render relevant MetaData entry which is of importance
The metadata entry (the score and comments) are only relevant in some
contexts.  It can be confusing which context something is relevant
in.  This change only renders the properties which you can enter when
showing the edit view.

There is a race condition when the views loads in which the wrong
component is rendered whilst the children are being loaded.  It is an
improvement in itself, so we'll accept that for now.