Polymorphic tree rendering under PointOfInterest
The tree is now rendered under PointOfInterest and uses polymorphism.
This change has created some gaps in the UX as the new UI is not
sufficiently consistent.  Although updates will likely need to follow,
the general design will probably be easier with this change.

As a follow-up there is a lot of dead code which should be treated.