Update to new routes and remove dead ones
We're updating the use to the new routes and we are removing the old
routes.  The routes we were using were based on the need to talk about
an experience.  That is not the case anymore.  The routes can now be
more logical, meaning we can convert to a bore sensical naming.