Support both PointOfInterest and Experience in path-instances
The path-instances utility creates intermediate classes for us in the
domain model.  It used to be that this always started from an
experience.  With the new inheritance setup, it's possible to start
this from another angle, being the PointOfInterest OR the Experience.
The new implementation checks the type of the supplied resource and
uses that to start.

It may be benificial to further abstract this so only classes are
being used, rather than magical names.  It's not obvious how we'd
drive this from the Ember Model definition and it's not obvious if
that is feasible.  We therefore let this rest for a while and intend
to revisit when further extending the functionality.