Remove edit-components/ prefix from custom-component-mapping
The custom component mapping has a prefix to identify the edit component
which should be rendered.  It is the intention to split off the logical
component from its rendering.  The prefix given here therefore doesn't
help overly much.

This will also make the widget more obvious as the edit-component name
can be removed there too.