Only show spinning icon when actually saving to backend
This feature only shows the fact that we're sending updates to the
backend when we're effectively doing so.  This could at least promote
the idea that the fast backend is not slow, even though it still implies
the same because the text says that we're saving.  It's possibly the
closest we'll get without showing the available guts of the backend.