Fix for saving choice items when no choice was set before
When no option was selected before, the choice criterion would not
correctly save the selection for the first option selected.
ember-data-change-tracker did not pick up the changed property and did
not send it to the backend.

We now save the item twice, as that seems to do the necessary magic.  It
is not apparent why this is the case and we should dive into it.