Cope with ember-data-change-tracker issue in manipulator
There seems to be an issue with ember-data-change-tracker when postponed
saving a relationship that has not been initialized yet.  This doesn't
seem to be saved and it's not immediately obvious why not.  Anyhow, a
workaround is to save the item directly in that case, which is what we
now do when we add files.