First step towards a mock login
We've noticed the mock login is not implemented as pretty as it could
be.  In this scope we aimed to implement a prettier mock login but ran
into issues with data refreshing.  As such we very much narrowed scope
to contain only part of the mock login and we'll add bits and pieces in
further commits.