A (barely) working state of the mock login
Introducing the mock login has occured in a more-or-less big-bang
solution in which we wired a bunch of things together and fixed issues
as they popped up.  We could probably split this up in multiple commits
but we'll likely split things incorrectly in that case.

As it stands with the current targets, we've chosen to keep this as it
is.