Implementation of transfer museum buttons
The museum can be transferred from the museum to inter, from inter to
the museum and from inter to public.  The buttons have been implemented
and they make the necessary call to the backend.  We currently assume
the transfer will be executed within time.  If it turns out that cannot
be achieved on the server, a different information mechanism should be
used with tasks.